code written in the EventsHighlight.tsx
