+++ src/app/Index.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Index = () => {
+  return (
+    <div>
+      helllo
+    </div>
+  )
+}
+
+export default Index
