+++ src/app/components/index.ts
@@ -1,7 +1,22 @@
+
+
 export { default as Header } from "./Header";
 export { default as HeroSection } from "./HeroSection";
-export { default as EventsHighlight } from "./EventsHighlight";
+// export {default as EventsHighlight} from "./EventsHighlight";
+export {EventsHighlight} from "./EventsHighlight";
 export { default as UpcomingEvents } from "./UpcomingEvents";
 export { default as Team } from "./Team";
 export { default as HighlightedProjects } from "./HighlightedProjects";
 export { default as Footer } from "./Footer";
+
+
+
+// {
+//   eventhighlight:function
+//   default: unknown
+// }
+
+
+
+
+
